+++ datadog_checks_dev/datadog_checks/dev/kube_port_forward.py
@@ -7,9 +7,9 @@ import os
 from contextlib import contextmanager
 
 from .env import environment_run
-from .ssh_tunnel import KillProcess, find_free_port, get_ip, run_background_command
+from .ssh_tunnel import KillProcess, run_background_command
 from .structures import LazyFunction, TempDir
-from .utils import chdir
+from .utils import chdir, find_free_port, get_ip
 
 PID_FILE = 'kubectl.pid'