+++ datadog_checks_dev/datadog_checks/dev/tooling/utils.py
@@ -11,8 +11,9 @@ import requests
 import semver
 from six import string_types
 
-from ..utils import file_exists, read_file
+from ..utils import file_exists, read_file, write_file
 from .constants import NOT_CHECKS, VERSION_BUMP, get_root
+from .git import get_latest_tag
 
 # match integration's version within the __about__.py module
 VERSION = re.compile(r'__version__ *= *(?:[\'"])(.+?)(?:[\'"])')
@@ -89,7 +90,19 @@ def get_config_spec(check_name):
     if check_name == 'agent':
         return os.path.join(get_root(), 'pkg', 'config', 'conf_spec.yaml')
     else:
-        return os.path.join(get_root(), check_name, 'datadog_checks', check_name, 'data', 'conf_spec.yaml')
+        path = load_manifest(check_name).get('assets', {}).get('configuration', {}).get('spec', '')
+        return os.path.join(get_root(), check_name, *path.split('/'))
+
+
+def get_default_config_spec(check_name):
+    return os.path.join(get_root(), check_name, 'assets', 'configuration', 'spec.yaml')
+
+
+def get_data_directory(check_name):
+    if check_name == 'agent':
+        return os.path.join(get_root(), 'pkg', 'config')
+    else:
+        return os.path.join(get_root(), check_name, 'datadog_checks', check_name, 'data')
 
 
 def get_config_files(check_name):
@@ -143,19 +156,29 @@ def get_version_string(check_name):
     """
     Get the version string for the given check.
     """
-    version = VERSION.search(read_version_file(check_name))
-    if version:
-        return version.group(1)
+    # Check the version file of the integration if available
+    # Otherwise, get the latest SemVer git tag for the project
+    if check_name:
+        version = VERSION.search(read_version_file(check_name))
+        if version:
+            return version.group(1)
+    else:
+        return get_latest_tag()
 
 
 def load_manifest(check_name):
     """
     Load the manifest file into a dictionary
     """
-    manifest_path = os.path.join(get_root(), check_name, 'manifest.json')
+    manifest_path = get_manifest_file(check_name)
     if file_exists(manifest_path):
-        return json.loads(read_file(manifest_path))
+        return json.loads(read_file(manifest_path).strip(), object_pairs_hook=OrderedDict)
     return {}
+
+
+def write_manifest(manifest, check_name):
+    manifest_path = get_manifest_file(check_name)
+    write_file(manifest_path, '{}\n'.format(json.dumps(manifest, indent=2)))
 
 
 def get_bump_function(changelog_types):