+++ ibm_db2/datadog_checks/ibm_db2/ibm_db2.py
@@ -12,7 +12,7 @@ from datadog_checks.base import AgentCheck, is_affirmative
 from datadog_checks.base.utils.containers import iter_unique
 
 from . import queries
-from .utils import scrub_connection_string, status_to_service_check
+from .utils import get_version, scrub_connection_string, status_to_service_check
 
 
 class IbmDb2Check(AgentCheck):
@@ -65,6 +65,40 @@ class IbmDb2Check(AgentCheck):
         self.query_table_space()
         self.query_transaction_log()
         self.query_custom()
+        self.collect_metadata()
+
+    def collect_metadata(self):
+        try:
+            raw_version = get_version(self._conn)
+        except Exception as e:
+            self.log.error("Error getting version: {}".format(str(e)))
+            return
+
+        if raw_version:
+            version_parts = self.parse_version(raw_version)
+            self.set_metadata('version', raw_version, scheme='parts', part_map=version_parts)
+
+            self.log.debug('Found ibm_db2 version: {}'.format(raw_version))
+        else:
+            self.log.warning('Could not retrieve ibm_db2 version info: {}'.format(raw_version))
+
+    def parse_version(self, version):
+        """
+        Raw version string is in format MM.mm.uuuu.
+        Parse version to MM.mm.xx.yy
+        where xx is the modification number and yy is the fix pack number
+        https://www.ibm.com/support/knowledgecenter/SSEPGG_11.1.0/com.ibm.db2.luw.wn.doc/doc/c0070229.html#c0070229
+        """
+        major, minor, update = version.split('.')
+        modification, fix = update[:2], update[2:]
+
+        # remove leading zeros from raw version parts
+        return {
+            'major': str(int(major)),
+            'minor': str(int(minor)),
+            'mod': str(int(modification)),
+            'fix': str(int(fix)),
+        }
 
     def query_instance(self):
         # Only 1 instance