+++ ibm_db2/tests/common.py
@@ -13,6 +13,7 @@ PORT = '50000'
 DB = 'datadog'
 USERNAME = 'db2inst1'
 PASSWORD = 'db2inst1-pwd'
+DB2_VERSION = os.getenv('DB2_VERSION')
 
 CONFIG = {
     'db': DB,